+++ forc-pkg/src/pkg.rs
@@ -1594,19 +1594,19 @@ pub fn dependency_namespace(
     node: NodeIx,
     engines: &Engines,
     contract_id_value: Option<ContractIdConst>,
-) -> Result<namespace::Module, vec1::Vec1<CompileError>> {
+) -> Result<namespace::Root, vec1::Vec1<CompileError>> {
     // TODO: Clean this up when config-time constants v1 are removed.
     let node_idx = &graph[node];
     let name = Some(Ident::new_no_span(node_idx.name.clone()));
-    let mut namespace = if let Some(contract_id_value) = contract_id_value {
+    let mut root_module = if let Some(contract_id_value) = contract_id_value {
         namespace::Module::default_with_contract_id(engines, name.clone(), contract_id_value)?
     } else {
         namespace::Module::default()
     };
 
-    namespace.is_external = true;
-    namespace.name = name;
-    namespace.visibility = Visibility::Public;
+    root_module.is_external = true;
+    root_module.name = name;
+    root_module.visibility = Visibility::Public;
 
     // Add direct dependencies.
     let mut core_added = false;
@@ -1640,7 +1640,7 @@ pub fn dependency_namespace(
                 ns
             }
         };
-        namespace.insert_submodule(dep_name, dep_namespace);
+        root_module.insert_submodule(dep_name, dep_namespace);
         let dep = &graph[dep_node];
         if dep.name == CORE {
             core_added = true;
@@ -1651,11 +1651,13 @@ pub fn dependency_namespace(
     if !core_added {
         if let Some(core_node) = find_core_dep(graph, node) {
             let core_namespace = &lib_namespace_map[&core_node];
-            namespace.insert_submodule(CORE.to_string(), core_namespace.clone());
+            root_module.insert_submodule(CORE.to_string(), core_namespace.clone());
         }
     }
 
-    let _ = namespace.star_import_with_reexports(
+    let mut root = namespace::Root::from(root_module);
+    
+    let _ = root.star_import_with_reexports(
         &Handler::default(),
         engines,
         &[CORE, PRELUDE].map(|s| Ident::new_no_span(s.into())),
@@ -1664,7 +1666,7 @@ pub fn dependency_namespace(
     );
 
     if has_std_dep(graph, node) {
-        let _ = namespace.star_import_with_reexports(
+        let _ = root.star_import_with_reexports(
             &Handler::default(),
             engines,
             &[STD, PRELUDE].map(|s| Ident::new_no_span(s.into())),
@@ -1673,7 +1675,7 @@ pub fn dependency_namespace(
         );
     }
 
-    Ok(namespace)
+    Ok(root)
 }
 
 /// Find the `std` dependency, if it is a direct one, of the given node.
@@ -1750,7 +1752,7 @@ pub fn compile(
     pkg: &PackageDescriptor,
     profile: &BuildProfile,
     engines: &Engines,
-    namespace: namespace::Module,
+    namespace: namespace::Root,
     source_map: &mut SourceMap,
 ) -> Result<CompiledPackage> {
     let mut metrics = PerformanceData::default();
